+++ lib/src/statistics_decimal.dart
@@ -1666,7 +1666,7 @@ extension DecimalOnIterableDecimalExtension on Iterable<Decimal> {
 
     var average = mean;
 
-    var first = itr.current;
+    var first = itr.current - average;
     var total = first * first;
 
     while (itr.moveNext()) {