+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.26
+
+- Fix:
+  - `IterableNumExtension.standardDeviation`.
+  - `IterableDoubleExtension.standardDeviation`.
+  - `DecimalOnIterableDecimalExtension.standardDeviation`.
+  - `DynamicIntOnIterableDynamicNumberExtension.standardDeviation`.
+
+- sdk: '>=2.14.0 <4.0.0'
+- intl: ^0.18.1
+- dependency_validator: ^3.2.3
+
 ## 1.0.25
 
 - `Decimal`: